+++ R/step_best_normalize.R
@@ -1,6 +1,8 @@
-#' bestNormalize transformation for \code{recipes} implementation
+#' Run bestNormalize transformation for \code{recipes} implementation
+#' 
+#' @aliases step_bestNormalize step_bestNormalize_new
 #'
-#' @description `step_bestNormalize` creates a specification of a recipe step
+#' @description `step_best_normalize` creates a specification of a recipe step
 #'   (see `recipes` package) that will transform data using the best of a suite
 #'   of normalization transformations estimated (by default) using
 #'   cross-validation.
@@ -31,7 +33,7 @@
 #'
 #' @details The bestnormalize transformation can be used to rescale a variable
 #'   to be more similar to a normal distribution. See `?bestNormalize` for more
-#'   information; `step_bestNormalize` is the implementation of `bestNormalize`
+#'   information; `step_best_normalize` is the implementation of `bestNormalize`
 #'   in the `recipes` context.
 #'
 #' @examples
@@ -39,7 +41,7 @@
 #' library(recipes)
 #' rec <- recipe(~ ., data = as.data.frame(iris))
 #'
-#' bn_trans <- step_bestNormalize(rec, all_numeric())
+#' bn_trans <- step_best_normalize(rec, all_numeric())
 #'
 #' bn_estimates <- prep(bn_trans, training = as.data.frame(iris))
 #'
@@ -56,7 +58,7 @@
 #'
 #' @importFrom recipes recipe rand_id add_step ellipse_check step
 #'   
-step_bestNormalize <-
+step_best_normalize <-
   function(recipe,
            ...,
            role = NA,
@@ -65,10 +67,10 @@ step_bestNormalize <-
            transform_options = list(),
            num_unique = 5,
            skip = FALSE,
-           id = rand_id("bestNormalize")) {
+           id = rand_id("best_normalize")) {
     add_step(
       recipe,
-      step_bestNormalize_new(
+      step_best_normalize_new(
         terms = ellipse_check(...),
         role = role,
         trained = trained,
@@ -81,10 +83,10 @@ step_bestNormalize <-
     )
   }
 
-step_bestNormalize_new <-
+step_best_normalize_new <-
   function(terms, role, trained, transform_info, transform_options, num_unique, skip, id) {
     step(
-      subclass = "bestNormalize",
+      subclass = "best_normalize",
       terms = terms,
       role = role,
       trained = trained,
@@ -98,7 +100,7 @@ step_bestNormalize_new <-
 
 #' @export
 #' @importFrom recipes prep terms_select check_type
-prep.step_bestNormalize <- function(x, training, info = NULL, ...) {
+prep.step_best_normalize <- function(x, training, info = NULL, ...) {
   col_names <- terms_select(x$terms, info = info)
   check_type(training[, col_names])
   
@@ -110,7 +112,7 @@ prep.step_bestNormalize <- function(x, training, info = NULL, ...) {
     num_unique = x$num_unique
   )
   
-  step_bestNormalize_new(
+  step_best_normalize_new(
     terms = x$terms,
     role = x$role,
     trained = TRUE,
@@ -125,7 +127,7 @@ prep.step_bestNormalize <- function(x, training, info = NULL, ...) {
 #' @export
 #' @importFrom recipes bake
 #' @importFrom tibble as_tibble
-bake.step_bestNormalize <- function(object, new_data, ...) {
+bake.step_best_normalize <- function(object, new_data, ...) {
   if (length(object$transform_info) == 0)
     return(as_tibble(new_data))
   param <- names(object$transform_info)
@@ -136,7 +138,7 @@ bake.step_bestNormalize <- function(object, new_data, ...) {
 }
 #' @export
 #' @importFrom recipes printer
-print.step_bestNormalize <-
+print.step_best_normalize <-
   function(x, width = max(20, options()$width - 35), ...) {
     cat("bestNormalize transformation on ", sep = "")
     printer(names(x$transform_info), x$terms, x$trained, width = width)
@@ -166,13 +168,13 @@ estimate_bn <- function(dat,
   res
 }
 
-#' @rdname step_bestNormalize
-#' @param x A `step_bestNormalize` object.
+#' @rdname step_best_normalize
+#' @param x A `step_best_normalize` object.
 #' @export
 #' @importFrom recipes tidy is_trained sel2char
 #' @importFrom tibble tibble
 #' 
-tidy.step_bestNormalize <- function(x, ...) {
+tidy.step_best_normalize <- function(x, ...) {
   if (is_trained(x)) {
     res <- tibble(terms = names(x$transform_info),
                   value = x$transform_info)
@@ -182,4 +184,16 @@ tidy.step_bestNormalize <- function(x, ...) {
   }
   res$id <- x$id
   res
+}
+
+#' @export
+step_bestNormalize <- function(...) {
+  .Deprecated("step_best_normalize", package = "bestNormalize", old = "step_bestNormalize") 
+  step_best_normalize(...)
+}
+
+#' @export
+step_bestNormalize_new <- function(...) {
+  .Deprecated("step_best_normalize_new", package = "bestNormalize", old = "step_bestNormalize_new") 
+  step_best_normalize_new(...)
 }